split `EnergyEvaluation` into multiple files
