+++ src/ParallelTemperingMonteCarlo.jl
@@ -7,7 +7,7 @@ include("BoundaryConditions.jl")
 include("Configurations.jl")
 include("Ensembles.jl")
 include("../MachineLearningPotential/MachineLearningPotential.jl")
-include("EnergyEvaluation.jl")
+include("energy_evaluation/EnergyEvaluation.jl")
 
 include("MCStates.jl")
 include("InputParams.jl")
